executor added

+++ internal/schedulerx/job/model/model.go
@@ -18,7 +18,7 @@ type Job struct {
 	Command   string `json:"command_id" gorm:"command;index:cmd_schedule_idx;uniqueIndex:cmd_schedule_unique"`
 	Status    string `json:"status" gorm:"status;default:'scheduled'"`
 	Schedule  int64  `json:"schedule" gorm:"schedule;index:cmd_schedule_idx;uniqueIndex:cmd_schedule_unique"`
-	StartedAt int    `json:"started_at" gorm:"started_at;default:0"`
-	EndedAt   int    `json:"ended_at" gorm:"ended_at;default:0"`
+	StartedAt int64  `json:"started_at" gorm:"started_at;default:0"`
+	EndedAt   int64  `json:"ended_at" gorm:"ended_at;default:0"`
 	Params    string `json:"params" gorm:"params;not null"`
 }
